changes a pais

+++ app/Http/Controllers/api/PaisController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\pais;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
 
 class PaisController extends Controller
 {
@@ -23,8 +24,10 @@ class PaisController extends Controller
      */
     public function store(Request $request)
     {
-        $pais = new pais();
+       $pais = new pais();
+        $pais->pais_codi = Str::upper($request->pais_codi);
         $pais->pais_nomb = $request->pais_nomb;
+        $pais->pais_capi = 0; 
         $pais->save();
 
         return json_encode(['pais' => $pais]);
@@ -44,15 +47,27 @@ class PaisController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        
         $pais = Pais::find($id);
-        $pais->pais_nomb = $request->name;
-        $pais->pais_capi = $request->capi;
+        if (!$pais) {
+            return response()->json(['message' => 'Pais not found'], 404);
+        }
+
+        $request->validate([
+            'pais_codi' => 'sometimes|string|size:3|unique:tb_pais,pais_codi,' . $id . ',pais_codi',
+            'pais_nomb' => 'sometimes|string|max:255',
+        ]);
+
+        if ($request->has('pais_codi')) {
+            $pais->pais_codi = strtoupper($request->pais_codi);
+        }
+        if ($request->has('pais_nomb')) {
+            $pais->pais_nomb = $request->pais_nomb;
+        }
+       
         $pais->save();
 
-        $pais = DB::table('tb_pais')
-            ->get();
-
-        return json_encode(['pais' => $pais]);
+        return response()->json(['pais' => $pais]);;
     }
 
     /**
@@ -61,10 +76,12 @@ class PaisController extends Controller
     public function destroy(string $id)
     {
         $pais = Pais::find($id);
+        if (!$pais) {
+            return response()->json(['message' => 'Pais not found'], 404);
+        }
         $pais->delete();
-        $pais = DB::table('tb_pais')
-            ->get();
+        $paises = DB::table('tb_pais')->get();
 
-        return json_encode(['paises' => $pais, 'success' => true]);
+        return response()->json(['paises' => $paises, 'success' => true]);
     }
 }
